# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: RFP Agent API
  version: 1.0.0
  description: |
    **RFP Agent** is an AI-powered platform for automated RFP discovery, proposal generation, and submission management.

    ## Features
    - 🔍 **Automated RFP Discovery**: Scan government portals for new opportunities
    - 🤖 **AI-Powered Proposal Generation**: Generate compliant proposals using GPT-5
    - 📊 **3-Tier Agent System**: Orchestrator, Managers, and Specialist agents
    - 📄 **Document Intelligence**: Parse and analyze RFP documents
    - 🚀 **Automated Submission**: Submit proposals through portal automation
    - 📈 **Analytics & Monitoring**: Track performance and success rates

    ## Architecture
    The platform uses a sophisticated 3-tier agentic system:
    - **Tier 1**: Primary Orchestrator (1 agent)
    - **Tier 2**: Managers - Portal, Proposal, Research (3 agents)
    - **Tier 3**: Specialists - Scanner, Monitor, Generator, Checker, Processor, Market Analyst, Historical Analyzer (7 agents)

    ## Authentication
    Currently, the API uses session-based authentication. API key support is planned for future releases.

  contact:
    name: RFP Agent Support
    email: support@rfpagent.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000/api
    description: Local development server
  - url: https://api.rfpagent.com/api
    description: Production server
...
paths:
...
  /rfps:
    get:
      tags: [RFPs]
      summary: List all RFPs
      description: Retrieve a paginated list of RFPs with optional filtering
      operationId: listRFPs
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [discovered, parsing, drafting, review, approved, submitted, closed]
          description: Filter by RFP status
        - name: portalId
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by portal ID
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
          description: Page number
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
          description: Items per page
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  rfps:
                    type: array
                    items:
                      $ref: '#/components/schemas/RFP'
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
              example:
                rfps:
                  - id: "550e8400-e29b-41d4-a716-446655440000"
                    title: "IT Services Procurement 2024"
                    agency: "City of Austin"
                    category: "Technology"
                    status: "discovered"
                    deadline: "2024-12-31T23:59:59Z"
                    estimatedValue: "500000.00"
                    sourceUrl: "https://financeonline.austintexas.gov/rfp/123"
                    progress: 20
                total: 150
                page: 1
                limit: 20
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags: [RFPs]
      summary: Create a new RFP
      description: Manually create a new RFP entry
      operationId: createRFP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRFPRequest'
            example:
              title: "Cloud Infrastructure Services"
              agency: "City of Philadelphia"
              category: "Technology"
              sourceUrl: "https://phlcontracts.phila.gov/contract/123456"
              deadline: "2024-12-31T23:59:59Z"
              estimatedValue: "750000.00"
      responses:
        '201':
          description: RFP created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RFP'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /rfps/{id}:
    get:
      tags: [RFPs]
      summary: Get RFP by ID
      description: Retrieve detailed information for a specific RFP
      operationId: getRFP
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: RFP ID
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RFP'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags: [RFPs]
      summary: Update an RFP
      description: Update RFP details
      operationId: updateRFP
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRFPRequest'
      responses:
        '200':
          description: RFP updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RFP'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags: [RFPs]
      summary: Delete an RFP
      description: Remove an RFP from the system
      operationId: deleteRFP
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: RFP deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /rfps/{id}/documents:
    get:
      tags: [RFPs, Documents]
      summary: Get RFP documents
      description: Retrieve all documents associated with an RFP
      operationId: getRFPDocuments
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Document'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags: [RFPs, Documents]
      summary: Add document to RFP
      description: Upload and associate a document with an RFP
      operationId: addRFPDocument
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - documentURL
                - filename
                - fileType
              properties:
                documentURL:
                  type: string
                  format: uri
                filename:
                  type: string
                fileType:
                  type: string
      responses:
        '201':
          description: Document added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  document:
                    $ref: '#/components/schemas/Document'
                  objectPath:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /rfps/manual:
    post:
      tags: [RFPs]
      summary: Process manual RFP submission
      description: Submit an RFP URL for automated scraping and processing
      operationId: submitManualRFP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - url
              properties:
                url:
                  type: string
                  format: uri
                  description: URL of the RFP page
                userNotes:
                  type: string
                  description: Optional notes about this RFP
            example:
              url: "https://financeonline.austintexas.gov/rfp/567890"
              userNotes: "High priority - cloud migration project"
      responses:
        '201':
          description: Manual RFP processing started
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  rfpId:
                    type: string
                    format: uuid
                  sessionId:
                    type: string
                  message:
                    type: string
              example:
                success: true
                rfpId: "550e8400-e29b-41d4-a716-446655440001"
                sessionId: "session_abc123"
                message: "RFP processing started successfully"
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ============================================================================
  # PROPOSAL ENDPOINTS
  # ============================================================================
  /proposals/rfp/{rfpId}:
    get:
      tags: [Proposals]
      summary: Get proposals for RFP
      description: Retrieve all proposals associated with a specific RFP
      operationId: getProposalsForRFP
      parameters:
        - name: rfpId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Proposal'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
...
    RFP:
      type: object
      required:
        - id
        - title
        - agency
        - sourceUrl
        - status
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        agency:
          type: string
        category:
          type: string
        portalId:
          type: string
          format: uuid
        sourceUrl:
          type: string
          format: uri
        deadline:
          type: string
          format: date-time
        estimatedValue:
          type: string
          pattern: '^\d+\.\d{2}$'
        status:
          type: string
          enum: [discovered, parsing, drafting, review, approved, submitted, closed]
        progress:
          type: integer
          minimum: 0
          maximum: 100
        requirements:
          type: object
        complianceItems:
          type: object
        riskFlags:
          type: object
        analysis:
          type: object
        addedBy:
          type: string
          enum: [manual, automatic]
        discoveredAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateRFPRequest:
      type: object
      required:
        - title
        - agency
        - sourceUrl
      properties:
        title:
          type: string
        description:
          type: string
        agency:
          type: string
        category:
          type: string
        portalId:
          type: string
          format: uuid
        sourceUrl:
          type: string
          format: uri
        deadline:
          type: string
          format: date-time
        estimatedValue:
          type: string
    UpdateRFPRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [discovered, parsing, drafting, review, approved, submitted, closed]
        progress:
          type: integer
          minimum: 0
          maximum: 100
        requirements:
          type: object
        complianceItems:
          type: object
        analysis:
          type: object
    Proposal:
      type: object
      required:
        - id
        - rfpId
        - status
      properties:
        id:
          type: string
          format: uuid
        rfpId:
          type: string
          format: uuid
        content:
          type: object
        narratives:
          type: object
        pricingTables:
          type: object
        forms:
          type: object
        attachments:
          type: object
        proposalData:
          type: object
        estimatedCost:
          type: string
        estimatedMargin:
          type: string
        status:
          type: string
          enum: [draft, review, approved, submitted]
        generatedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Document:
      type: object
      required:
        - id
        - rfpId
        - filename
        - fileType
        - objectPath
      properties:
        id:
          type: string
          format: uuid
        rfpId:
          type: string
          format: uuid
        filename:
          type: string
        fileType:
          type: string
        objectPath:
          type: string
        extractedText:
          type: string
        parsedData:
          type: object
        uploadedAt:
          type: string
          format: date-time
...
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              details:
                type: array
                items:
                  type: object
          example:
            error: "Invalid input data"
            details:
              - field: "url"
                message: "Must be a valid URL"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
          example:
            error: "Resource not found"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
          example:
            error: "Internal server error"

  # ============================================================================
  # SECURITY SCHEMES
  # ============================================================================
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: connect.sid
      description: Session-based authentication using cookies

    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication (planned for future release)

# Global security - currently using session-based auth
security:
  - sessionAuth: []